+++ R/03-modelResults2DKernel.R
@@ -18,6 +18,9 @@ modelResults2DKernelUI <- function(id, title = "", asFruitsTab = FALSE){
       sidebarPanel(
         width = 2,
         style = "position:fixed; width:14%; max-width:220px; overflow-y:auto; height:88%",
+        downUploadButtonUI(ns("downUpload"), title = "Load a Model", label = "Upload / Download"),
+        textAreaInput(ns("modelNotes"), label = NULL, placeholder = "Description ..."),
+        tags$hr(),
         selectInput(ns("dataSource"),
                     "Data source",
                     choices = if (!asFruitsTab) c("Database" = "db", "Upload file" = "file", "Saved map" = "model") else c("Database" = "db"),
@@ -57,7 +60,7 @@ modelResults2DKernelUI <- function(id, title = "", asFruitsTab = FALSE){
           selectInput(inputId = ns("Latitude"),
                       label = "Latitude variable:",
                       choices = c("Latitude")),
-          selectInput(inputId = ns("Independent"),
+          selectInput(inputId = ns("IndependentX"),
                       label = "Presence/Absence variable (optional):",
                       choices = c("")),
           selectInput(inputId = ns("Weighting"),
@@ -95,17 +98,19 @@ modelResults2DKernelUI <- function(id, title = "", asFruitsTab = FALSE){
           min = 5, max = 100, step = 1, value = 10),
           conditionalPanel(
             condition = "input.modelArea == true",
+            tags$strong("Latitude restriction:"),
             numericInput(inputId = ns("mALat1"),
-                         label = "Set lower latitude restriction",
+                         label = "Lower",
                          min = -90, max = 90, value = c(-90), width = "80%"),
             numericInput(inputId = ns("mALat2"),
-                         label = "Set upper latitude restriction",
+                         label = "Upper",
                          min = -90, max = 90, value = c(90), width = "80%"),
+            tags$strong("Longitude restriction:"),
             numericInput(inputId = ns("mALong1"),
-                         label = "Set lower longitude restriction",
+                         label = "Lower",
                          min = -180, max = 180, value = c(-180), width = "80%"),
             numericInput(inputId = ns("mALong2"),
-                         label = "Set upper longitude restriction",
+                         label = "Upper",
                          min = -180, max = 180, value = c(180), width = "80%"),
             ns = ns
           )
@@ -405,12 +410,15 @@ modelResults2DKernel <- function(input, output, session, isoData, savedMaps, fru
     updateTextInput(session, "saveMapName", value = "")
   })
 
-  data <- reactive({
-    switch(
+  data <- reactiveVal()
+  observe({
+    activeData <- switch(
       input$dataSource,
       db = isoData(),
       file = fileImport()
     )
+
+    data(activeData)
   })
 
   coordType <- reactive({
@@ -423,6 +431,46 @@ modelResults2DKernel <- function(input, output, session, isoData, savedMaps, fru
 
   Model <- reactiveVal(NULL)
 
+  # MODEL DOWN- / UPLOAD ----
+  uploadedData <- downUploadButtonServer(
+    "downUpload",
+    dat = data,
+    inputs = input,
+    model = Model,
+    rPackageName = "MpiIsoApp",
+    githubRepo = "iso-app",
+    subFolder = "KernelR",
+    helpHTML = getHelp(id = "model2DKernel"),
+    modelNotes = reactive(input$modelNotes),
+    compressionLevel = 1)
+
+  observe(priority = 100, {
+    ## update data ----
+    data(uploadedData$data)
+  }) %>%
+    bindEvent(uploadedData$data)
+
+  observe(priority = 50, {
+    ## reset input of model notes
+    updateTextAreaInput(session, "modelNotes", value = "")
+
+    ## update inputs ----
+    inputIDs <- names(uploadedData$inputs)
+    inputIDs <- inputIDs[inputIDs %in% names(input)]
+
+    for (i in 1:length(inputIDs)) {
+      session$sendInputMessage(inputIDs[i],  list(value = uploadedData$inputs[[inputIDs[i]]]) )
+    }
+  }) %>%
+    bindEvent(uploadedData$inputs)
+
+  observe(priority = 10, {
+    ## update model ----
+    Model(uploadedData$model)
+  }) %>%
+    bindEvent(uploadedData$model)
+
+  # RUN MODEL ----
   observeEvent(input$start, ignoreNULL = FALSE, {
     if (input$dataSource == "model") {
       if (length(savedMaps()) == 0) return(NULL)
@@ -445,7 +493,7 @@ modelResults2DKernel <- function(input, output, session, isoData, savedMaps, fru
     data <- data()
 
     model <- withProgress(
-      estimateMapKernel(data = data, independent = input$Independent,
+      estimateMapKernel(data = data, independent = input$IndependentX,
                   Longitude = input$Longitude, Latitude = input$Latitude,
                   CoordType = coordType(),
                   Weighting = input$Weighting,
@@ -763,7 +811,9 @@ modelResults2DKernel <- function(input, output, session, isoData, savedMaps, fru
 
   output$DistMap <- renderPlot({
     validate(validInput(Model()))
-    res <- plotFun()(Model())
+    withProgress({
+      res <- plotFun()(Model())
+    }, min = 0, max = 1, value = 0.8, message = "Plotting map ...")
     values$predictions <- res$XPred
     values$meanCenter <- res$meanCenter
     values$sdCenter <- res$sdCenter
@@ -787,7 +837,7 @@ modelResults2DKernel <- function(input, output, session, isoData, savedMaps, fru
     centerEstimate$text()
   })
 
-  observe({
+  observe(priority = 75, {
     numVars <- unlist(lapply(names(data()), function(x){
       if (
         (is.integer(data()[[x]]) | is.numeric(data()[[x]]) | sum(!is.na(as.numeric((data()[[x]])))) > 2) #&
@@ -818,7 +868,7 @@ modelResults2DKernel <- function(input, output, session, isoData, savedMaps, fru
       selectedLatitude <- "latitude"
     }
 
-    updateSelectInput(session, "Independent", choices = c("", numVars),
+    updateSelectInput(session, "IndependentX",  choices = c("", numVars),
                       selected = selectedIndependent)
     updateSelectInput(session, "Weighting", choices = c("", numVars))
     updateSelectInput(session, "Longitude", choices = c("", names(data())),
@@ -831,8 +881,8 @@ modelResults2DKernel <- function(input, output, session, isoData, savedMaps, fru
                       selected = selectedTextLabel)
     updateSelectInput(session, "pointLabelsVarCol", choices = c("", names(data())),
                       selected = selectedTextLabel)
-
-  })
+  }) %>%
+    bindEvent(data())
 
   ## Import Data ----
   importedDat <- importDataServer("importData")