+++ R/03-modelResultsSpread.R
@@ -17,6 +17,9 @@ modelResultsSpreadUI <- function(id, title = ""){
       sidebarPanel(
         width = 2,
         style = "position:fixed; width:14%; max-width:220px; overflow-y:auto; height:88%",
+        downUploadButtonUI(ns("downUpload"), title = "Load a Model", label = "Upload / Download"),
+        textAreaInput(ns("modelNotes"), label = NULL, placeholder = "Description ..."),
+        tags$hr(),
         selectInput(ns("dataSource"),
                     "Data source",
                     choices = c("Database" = "db",
@@ -108,9 +111,6 @@ modelResultsSpreadUI <- function(id, title = ""){
           checkboxInput(inputId = ns("Outlier"),
                         label = "Remove model outliers",
                         value = FALSE, width = "100%"),
-          checkboxInput(inputId = ns("OutlierD"),
-                        label = "Remove data outliers",
-                        value = FALSE, width = "100%"),
           conditionalPanel(
             condition = "input.Outlier == true",
             sliderInput(inputId = ns("OutlierValue"),
@@ -118,6 +118,9 @@ modelResultsSpreadUI <- function(id, title = ""){
                         min = 2, max = 8, value = 4, step = 0.1),
             ns = ns
           ),
+          checkboxInput(inputId = ns("OutlierD"),
+                        label = "Remove data outliers",
+                        value = FALSE, width = "100%"),
           conditionalPanel(
             condition = "input.OutlierD == true",
             sliderInput(inputId = ns("OutlierValueD"),
@@ -130,17 +133,19 @@ modelResultsSpreadUI <- function(id, title = ""){
                         value = FALSE, width = "100%"),
           conditionalPanel(
             condition = "input.modelArea == true",
+            tags$strong("Latitude restriction:"),
             numericInput(inputId = ns("mALat1"),
-                         label = "Set lower latitude restriction",
+                         label = "Lower",
                          min = -90, max = 90, value = c(-90), width = "80%"),
             numericInput(inputId = ns("mALat2"),
-                         label = "Set upper latitude restriction",
+                         label = "Upper",
                          min = -90, max = 90, value = c(90), width = "80%"),
+            tags$strong("Longitude restriction:"),
             numericInput(inputId = ns("mALong1"),
-                         label = "Set lower longitude restriction",
+                         label = "Lower",
                          min = -180, max = 180, value = c(-180), width = "80%"),
             numericInput(inputId = ns("mALong2"),
-                         label = "Set upper longitude restriction",
+                         label = "Upper",
                          min = -180, max = 180, value = c(180), width = "80%"),
             ns = ns
           ),
@@ -443,13 +448,15 @@ modelResultsSpread <- function(input, output, session, isoData, savedMaps, fruit
     updateTextInput(session, "saveMapName", value = "")
   })
 
-
-  data <- reactive({
-    switch(
+  data <- reactiveVal()
+  observe({
+    activeData <- switch(
       input$dataSource,
       db = isoData(),
       file = fileImport()
     )
+
+    data(activeData)
   })
 
   coordType <- reactive({
@@ -462,6 +469,46 @@ modelResultsSpread <- function(input, output, session, isoData, savedMaps, fruit
 
   Model <- reactiveVal()
 
+  # MODEL DOWN- / UPLOAD ----
+  uploadedData <- downUploadButtonServer(
+    "downUpload",
+    dat = data,
+    inputs = input,
+    model = Model,
+    rPackageName = "MpiIsoApp",
+    githubRepo = "iso-app",
+    subFolder = "SpreadR",
+    helpHTML = getHelp(id = "spread"),
+    modelNotes = reactive(input$modelNotes),
+    compressionLevel = 1)
+
+  observe(priority = 100, {
+    ## update data ----
+    data(uploadedData$data)
+  }) %>%
+    bindEvent(uploadedData$data)
+
+  observe(priority = 50, {
+    ## reset input of model notes
+    updateTextAreaInput(session, "modelNotes", value = "")
+
+    ## update inputs ----
+    inputIDs <- names(uploadedData$inputs)
+    inputIDs <- inputIDs[inputIDs %in% names(input)]
+
+    for (i in 1:length(inputIDs)) {
+      session$sendInputMessage(inputIDs[i],  list(value = uploadedData$inputs[[inputIDs[i]]]) )
+    }
+  }) %>%
+    bindEvent(uploadedData$inputs)
+
+  observe(priority = 10, {
+    ## update model ----
+    Model(uploadedData$model)
+  }) %>%
+    bindEvent(uploadedData$model)
+
+  # RUN MODEL ----
   observeEvent(input$start, ignoreNULL = FALSE, {
     if (input$dataSource == "model") {
       if (length(savedMaps()) == 0) return(NULL)
@@ -488,7 +535,7 @@ modelResultsSpread <- function(input, output, session, isoData, savedMaps, fruit
 
   Independent <- reactive({
     if (input$dataSource == "model") names(Model()$data)[1]
-    else input$Independent
+    else input$IndependentX
   })
 
   zoomFromModel <- reactiveVal(50)
@@ -818,7 +865,9 @@ modelResultsSpread <- function(input, output, session, isoData, savedMaps, fruit
 
   output$DistMap <- renderPlot({
     validate(validInput(Model()))
-    res <- plotFun()(Model())
+    withProgress({
+      res <- plotFun()(Model())
+    }, min = 0, max = 1, value = 0.8, message = "Plotting map ...")
     values$predictions <- res$XPred
     values$meanCenter <- res$meanCenter
     values$sdCenter <- res$sdCenter
@@ -835,7 +884,7 @@ modelResultsSpread <- function(input, output, session, isoData, savedMaps, fruit
     centerEstimate$text()
   })
 
-  observe({
+  observe(priority = 75, {
     numVars <- unlist(lapply(names(data()), function(x){
       if (
         (is.integer(data()[[x]]) | is.numeric(data()[[x]]) | sum(!is.na(as.numeric((data()[[x]])))) > 3) #&
@@ -881,7 +930,8 @@ modelResultsSpread <- function(input, output, session, isoData, savedMaps, fruit
       updateSelectInput(session, "DateOne", choices = c("", numVars))
       updateSelectInput(session, "DateTwo", choices = c("", numVars))
   #   }
-   })
+   }) %>%
+    bindEvent(data())
 
   ## Import Data ----
   importedDat <- importDataServer("importData")